add point controller

+++ src/main/java/com/nhnacademy/taskapi/point/domain/Point.java
@@ -1,14 +1,12 @@
 package com.nhnacademy.taskapi.point.domain;
 
-import com.nhnacademy.taskapi.Member;
+import com.nhnacademy.taskapi.member.domain.Member;
 import jakarta.persistence.*;
 import jakarta.validation.constraints.NotNull;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-
-import java.math.BigDecimal;
 
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
@@ -21,19 +19,19 @@ public class Point {
 
     @NotNull
     @Column(nullable = false, precision = 10, scale = 2)
-    private BigDecimal pointCurrent;
+    private int pointCurrent;
 
     @ManyToOne(optional = false, fetch = FetchType.LAZY)
     @JoinColumn(nullable = false, name = "member_id")
     private Member member;
 
     @Builder
-    public Point(BigDecimal pointCurrent, Member member) {
+    public Point(int pointCurrent, Member member) {
         this.pointCurrent = pointCurrent;
         this.member = member;
     }
 
-    public void updatePointCurrent(BigDecimal pointCurrent) {
+    public void updatePointCurrent(int pointCurrent) {
         this.pointCurrent = pointCurrent;
     }
 }
